Error dependecis according to a format and in coommon
